Move FormattedReporter to components and apply to bulk import and sighting headers
