+++ src/pages/assetGroup/AssetGroup.jsx
@@ -16,8 +16,7 @@ import queryKeys from '../../constants/queryKeys';
 import MainColumn from '../../components/MainColumn';
 import LoadingScreen from '../../components/LoadingScreen';
 import SadScreen from '../../components/SadScreen';
-import Text from '../../components/Text';
-import Link from '../../components/Link';
+import FormattedReporter from '../../components/formatters/FormattedReporter';
 import MoreMenu from '../../components/MoreMenu';
 import ConfirmDelete from '../../components/ConfirmDelete';
 import EntityHeader from '../../components/EntityHeader';
@@ -106,10 +105,6 @@ export default function AssetGroup() {
   const dateCreated = get(data, 'created');
 
   const sightingCreator = data?.creator;
-  const creatorName =
-    sightingCreator?.full_name ||
-    intl.formatMessage({ id: 'UNNAMED_USER' });
-  const creatorUrl = `/users/${sightingCreator?.guid}`;
 
   const pipelineStatusPreparation = get(
     data,
@@ -167,10 +162,13 @@ export default function AssetGroup() {
         }
       >
         {sightingCreator && (
-          <Text variant="body2">
-            {intl.formatMessage({ id: 'REPORTED_BY' })}
-            <Link to={creatorUrl}>{creatorName}</Link>
-          </Text>
+          <FormattedReporter
+            variant="body2"
+            reporter={{
+              guid: sightingCreator.guid,
+              fullName: sightingCreator.full_name,
+            }}
+          />
         )}
       </EntityHeader>
       {isPreparationFailed && (
